[IMP] improve views,setting,code
