+++ limit_validation_jsi/models/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-
+from . import res_config_settings
 from . import sale_order
 from . import stock_picking
