+++ limit_validation_jsi/models/stock_picking.py
@@ -1,7 +1,5 @@
 # -*- coding: utf-8 -*-
 
-import datetime
-
 from odoo import api, fields, models, _
