+++ limit_validation_jsi/views/sale_order_views.xml
@@ -6,7 +6,7 @@
             <field name="model">sale.order</field>
             <field name="inherit_id" ref="sale.view_order_form"/>
             <field name="arch" type="xml">
-                <field name="date_order" position="after">
+                <field name="carrier_id" position="after">
                     <field name="is_overdue_limit_end"/>
                 </field>
                 <xpath expr="//button[@id='action_confirm']" position="attributes">
